WIP forum markup tag helper
This should decrease the amount of extra copying being done.
