+++ TASVideos/Pages/BasePageModel.cs
@@ -37,7 +37,9 @@ namespace TASVideos.Pages
 		{
 			var parsed = PostParser.Parse(text, useBbCode, useHtml);
 			using var writer = new StringWriter();
+			writer.Write("<div class=postbody>");
 			parsed.WriteHtml(writer);
+			writer.Write("</div>");
 			return writer.ToString();
 		}
